lib: Add qcow_stream module to decode a QCOW2 file from a stream
Qcow library code in general and its dependencies assume that the input
and output are seekable files. It is often useful to be able to decode a
qcow2 file from a non-seekable stream instead (in xen-api's case, when
importing a virtual disk we can't create an intermediate Qcow2 file, we
want to be able to start creating the disk on-the-fly).

Provide a Qcow_stream module that implements a streaming decode
(qcow2-to-raw) functionality. Upstream QEMU already offers a raw-to-qcow2-stream
script that does the inverse:
https://github.com/qemu/qemu/commit/df957115c46845e2c0ccc29ac0a75eb9700a9a0d

Qcow_stream largely follows the implementation of the decode functionality
offered in cli/impl.ml, reimplementing some things in a non-seekable manner
where required.

Separate it out into its own library with minimal dependencies, so that
Qcow_stream can be used without being dependent on Qcow and its
dependencies.

Remove mirage-block-unix from the dependencies in lib/dune, it was in
fact only required in cli/dune and used there transitively.

Signed-off-by: Andrii Sultanov <andriy.sultanov@vates.tech>